pyproject.toml
[project]
name = "backend-plataforma-ganep-lar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firebase-admin>=6.6.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "google-api-python-client>=2.162.0",
    "gspread>=6.1.4",
    "oauth2client>=4.1.3",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
]
